Uploaded our Problem Statement
The problem statement is derived from actual possible case studies for different specialized requirement needs of the user. All the possible obstacles faced or to be facing must be predicted and solutions must be obtained during the planning and developing of the project. These statements are the key to identify the actual user requirement. Giving it in the form of case study enhances the priority levels of achieving the expected outcome.